requirements.txt
a2wsgi==1.7.0
bidict==0.22.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-engineio==4.5.1
python-socketio==5.8.0
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
